requirements.txt
attrs==23.1.0
aws-cdk-lib==2.78.0
aws-cdk.asset-awscli-v1==2.2.198
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.165
aws-cdk.aws-glue-alpha==2.78.0a0
cattrs==23.1.2
constructs==10.2.55
importlib-resources==5.12.0
jsii==1.84.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions===4.4.0
